@@ -6,24 +6,24 @@
     "@mdi/js": "^4.9.95",
     "@mdi/react": "^1.4.0",
     "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
+    "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.1.2",
     "@types/react-transition-group": "^4.2.4",
     "axios": "^0.19.2",
     "bootstrap": "^4.4.1",
-    "camelcase-keys": "^6.1.2",
-    "capitalize": "^2.0.2",
+    "camelcase-keys": "^6.2.2",
+    "capitalize": "^2.0.3",
     "chart.js": "^2.9.3",
     "cuid": "^2.1.8",
-    "http-proxy-middleware": "^1.0.1",
+    "http-proxy-middleware": "^1.0.3",
     "json-form-data": "^1.7.2",
     "kind-of": "^6.0.3",
     "lodash": "^4.17.15",
     "node-sass": "^4.13.1",
     "pluralize": "^8.0.0",
-    "react": "^16.12.0",
+    "react": "^16.13.1",
     "react-chartjs-2": "^2.9.0",
-    "react-dom": "^16.12.0",
+    "react-dom": "^16.13.1",
     "react-height": "^3.0.1",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
@@ -32,7 +32,7 @@
     "react-uuid": "^1.0.2",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
-    "snakecase-keys": "^3.1.2"
+    "snakecase-keys": "^3.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -56,6 +56,6 @@
     ]
   },
   "devDependencies": {
-    "css-loader": "^3.4.2"
+    "css-loader": "^3.5.2"
   }
 }
